Rock, Paper, Scissors
This code is the result of a tutorial from the book 'Sinatra, Up and Running' by Alan
Harris & Konstantin Haase.

In the beginning, it was really the same as in the book, but it didn't work on my machine.

Things that I needed to change, before it worked:
- add require rubygems as well as require backports. The later was required, because the
tutorial is for Ruby 1.9 but I'm still using 1.8
- the Hash
Actually I don't know. May be it would work now the way it is in the book, after I added require
backports, but before I got the stranges error messages like: "odd number of arguments for hash".
Well, I know that 3 is an odd number, but what does this have to do with my hash? When I removed,
or added an other Element I used to get another error.

But: It works now! :D

	new file:   .gitignore
	new file:   game.rb